requirements.txt
pysolr==3.8.1
emoji==0.5.3
regex==2019.8.19
matplotlib==3.1.0
tqdm 
nltk 
numpy 
segtok==1.5.7
flair==0.6.1
pytorch-transformers==1.2.0
transformers==4.20.1
colorama==0.4.3
termcolor==1.1.0
faiss-cpu>=1.6.1
prettytable
pandas
sentence-transformers
wandb
pytrec-eval
openai
opencv-python
easyocr
imutils
selenium
webdriver-manager